requirements.txt
Django
djangorestframework 
python-decouple
django-cors-headers
requests